# Optimize assertion rewriting